+++ app/controllers/staff/activities_controller.rb
@@ -17,6 +17,7 @@ class Staff::ActivitiesController < Staff::BaseController
         organisation: @organisation,
         scope: :current
       ).call
+      add_breadcrumb t("page_content.breadcrumbs.current_index"), organisation_activities_path(@organisation)
     end
   end
 
@@ -53,6 +54,7 @@ class Staff::ActivitiesController < Staff::BaseController
         organisation: @organisation,
         scope: :historic
       ).call
+      add_breadcrumb t("page_content.breadcrumbs.historic_index"), historic_organisation_activities_path(@organisation)
     end
   end