+++ app/controllers/staff/base_controller.rb
@@ -6,4 +6,6 @@ class Staff::BaseController < ApplicationController
   # Ensure that Pundit 'authorize' and scopes are used
   after_action :verify_authorized, except: :index
   after_action :verify_policy_scoped, only: :index
+
+  add_breadcrumb "Home", :root_path
 end