+++ app/helpers/activity_helper.rb
@@ -1,4 +1,6 @@
 module ActivityHelper
+  include CodelistHelper
+
   def step_is_complete_or_next?(activity:, step:)
     steps = Activity::FORM_STEPS
 
@@ -32,5 +34,23 @@ module ActivityHelper
 
   def policy_markers_desertification_iati_codes_to_enum(code)
     Activity::DESERTIFICATION_POLICY_MARKER_CODES.key(code.to_i)
+  end
+
+  def benefitting_countries_with_percentages(benefitting_countries)
+    return [] if benefitting_countries.blank?
+
+    # Get an equal percentage split between all the countries
+    # (together with the remainder if possible)
+    percentage, remainder = 100.divmod(benefitting_countries.count)
+
+    benefitting_countries.map do |country|
+      # If we're at the last item, add the remainder to the percentage
+      # split
+      percentage += remainder if country == benefitting_countries.last
+
+      OpenStruct.new(code: country,
+                     name: country_name_from_code(country),
+                     percentage: percentage.to_f)
+    end
   end
 end