+++ spec/features/staff/users_can_view_an_activity_as_xml_spec.rb
@@ -32,6 +32,50 @@ RSpec.feature "Users can view an activity as XML" do
         end
       end
 
+      context "when the activity has benefitting countries" do
+        let(:activity) {
+          create(:fund_activity,
+            organisation: organisation,
+            delivery_partner_identifier: "IND-ENT-IFIER",
+            benefitting_countries: ["CV"])
+        }
+        let(:xml) { Nokogiri::XML::Document.parse(page.body) }
+
+        it "contains data about the recipient country" do
+          visit organisation_activity_path(organisation, activity, format: :xml)
+
+          expect(xml.at("iati-activity/recipient-country/@code").text).to eq(activity.benefitting_countries.first)
+          expect(xml.at("iati-activity/recipient-country/@percentage").text).to eq("100.0")
+          expect(xml.at("iati-activity/recipient-country/narrative").text).to eq("Cabo Verde")
+        end
+      end
+
+      context "when the activity has multiple benefitting countries" do
+        let(:activity) {
+          create(:fund_activity,
+            organisation: organisation,
+            delivery_partner_identifier: "IND-ENT-IFIER",
+            benefitting_countries: ["CV", "BZ"])
+        }
+        let(:xml) { Nokogiri::XML::Document.parse(page.body) }
+
+        it "contains data about the recipient country" do
+          visit organisation_activity_path(organisation, activity, format: :xml)
+
+          results = xml.xpath("//iati-activity/recipient-country")
+
+          expect(results.count).to eq(2)
+
+          expect(results[0].at("@code").text).to eq("CV")
+          expect(results[0].at("@percentage").text).to eq("50.0")
+          expect(results[0].at("narrative").text).to eq("Cabo Verde")
+
+          expect(results[1].at("@code").text).to eq("BZ")
+          expect(results[1].at("@percentage").text).to eq("50.0")
+          expect(results[1].at("narrative").text).to eq("Belize")
+        end
+      end
+
       context "when the activity has recipient_region geography" do
         let(:activity) {
           create(:fund_activity,